requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
watchtower==3.0.1
boto3==1.28.38
pytz==2023.3
matplotlib==3.7.2
numpy==1.25.2
jsonschema==4.19.0
typing-extensions==4.7.1
